decode starting to work
