+++ pg_analytics/src/rmgr/mod.rs
@@ -1,14 +1,13 @@
-mod desc;
 pub mod xlog;
 
 use once_cell::sync::Lazy;
 use pgrx::pg_sys::AsPgCStr;
 use pgrx::*;
+use shared::postgres::wal::{xlog_rec_get_data, xlog_rec_get_info};
 
-use crate::rmgr::desc::rm_desc;
-use crate::rmgr::xlog::XLOG_INSERT;
+use crate::rmgr::xlog::{XLogInsertRecord, XLOG_INSERT};
 
-pub static CUSTOM_RMGR_ID: u8 = 135;
+pub static CUSTOM_RMGR_ID: u8 = 128;
 
 pub static mut CUSTOM_RMGR: Lazy<pg_sys::RmgrData> = Lazy::new(|| pg_sys::RmgrData {
     rm_name: "pg_analytics".as_pg_cstr(),
@@ -23,7 +22,28 @@ pub static mut CUSTOM_RMGR: Lazy<pg_sys::RmgrData> = Lazy::new(|| pg_sys::RmgrDa
 
 unsafe extern "C" fn rm_startup() {}
 unsafe extern "C" fn rm_cleanup() {}
-unsafe extern "C" fn rm_redo(record: *mut pg_sys::XLogReaderState) {}
+
+pub unsafe extern "C" fn rm_desc(
+    buf: *mut pg_sys::StringInfoData,
+    record: *mut pg_sys::XLogReaderState,
+) {
+    let decoded_record = xlog_rec_get_data(record) as *mut XLogInsertRecord;
+    let info_mask = pg_sys::XLR_INFO_MASK as u8;
+    let info = xlog_rec_get_info(record) & !info_mask;
+
+    if info == XLOG_INSERT {
+        pg_sys::appendStringInfo(buf, "flags: insert".as_pg_cstr());
+    }
+}
+
+unsafe extern "C" fn rm_redo(record: *mut pg_sys::XLogReaderState) {
+    ereport!(
+        PgLogLevel::LOG,
+        PgSqlErrorCode::ERRCODE_SUCCESSFUL_COMPLETION,
+        "FROM REDO"
+    );
+}
+
 unsafe extern "C" fn rm_mask(page_data: *mut i8, block_number: u32) {}
 unsafe extern "C" fn rm_decode(
     context: *mut pg_sys::LogicalDecodingContext,
@@ -33,7 +53,7 @@ unsafe extern "C" fn rm_decode(
 
 unsafe extern "C" fn rm_identify(info: u8) -> *const i8 {
     let info_mask = pg_sys::XLR_INFO_MASK as u8;
-    if (info & !info_mask == XLOG_INSERT) {
+    if (info & !info_mask) == XLOG_INSERT {
         "INSERT".as_pg_cstr()
     } else {
         "UNKNOWN".as_pg_cstr()
